Add hidden boolean
Add tests to check if a hide_contact boolean
changes state. Book suggests implementing if one
feels like it so I did. There was also a typo in
the example test:

```
# original
expect(@contact.reload.hidden?).to be_true

# correct
expect(@contact.reload.hidden?).to be true
```

as `true` would not `respond to true?`