Test nested route
For this I’ve just scaffolded the phone controller
as a quick and dirty way to test the very short
section in the chapter.

The auto-generated test looks useful to look at so
have commented out all of the pending tests for
now to keep the test output clearer for now.